Addresses vehicle vibration issue #1
Mitigated abnormal vehicle vibration of issue #1 by adjusting CollisionShape dimensions. The shape seems to significantly impact the strength of vehicle vibrations in Godot. This depends primarily on whether the CollisionShape length extends beyond the wheel placement, followed by a marginal significance of ColllisionShape width, regardless of wheel placement. CollisionShape depth is seemingly unimportant.

The shape also seems to affect whether the vibration is introduced to situations of braking to a stop, and continuing to apply the brake when stopped.

Changing wheel_friction_slip could reduce vibrations, but doing so directly impairs handling.